Move `lookup` utility function
